+++ codecs/rubixDataEncoding/decoder_test.go
@@ -6,8 +6,7 @@ import (
 )
 
 func testHeader(t *testing.T, serialData *SerialData, expectedHeader MetaDataKey) {
-	var pos uint8
-	header := getMetaDataKey(serialData, &pos)
+	header, _ := parseMetaData(serialData)
 	if header != expectedHeader {
 		t.Fatalf("Expected header %v, but got %v", expectedHeader, header)
 	}
@@ -143,7 +142,7 @@ func TestDecodefromStm32Encode(t *testing.T) {
 	testDecodeFloat(t, &serialData, 333, MDK_CO2)
 	testDecodeFloat(t, &serialData, 2.9, MDK_BATTERY_VOLTAGE)
 	testDecodeFloat(t, &serialData, 15, MDK_PUSH_FREQUENCY)
-	testDecodeFloat(t, &serialData, 0.888, MDK_RAW)
+	testDecodeFloat(t, &serialData, 0.888, MDK_ANALOG_IN)
 	testDecodeFloat(t, &serialData, 22, MDK_FIRMWARE_VERSION)
 	testDecodeFloat(t, &serialData, 44, MDK_HARDWARE_VERSION)
 	testDecodeUint(t, &serialData, (uint8)(1), MDK_UINT_8)
@@ -202,17 +201,17 @@ func TestSerialDataRaw(t *testing.T) {
 
 	Require(t, canDecode(serialData))
 
-	testDecodeFloat(t, serialData, 0.1, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.2, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.3, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.4, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.5, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.1, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.2, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.3, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.4, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.5, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.6, MDK_RAW)
+	testDecodeFloat(t, serialData, 0.1, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.2, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.3, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.4, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.5, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.1, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.2, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.3, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.4, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.5, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.6, MDK_ANALOG_IN)
 	testDecodeFloat(t, serialData, 1, MDK_DIGITAL)
 	testDecodeFloat(t, serialData, 1, MDK_DIGITAL)
 
@@ -246,42 +245,9 @@ func TestPositional(t *testing.T) {
 
 	Require(t, canDecode(serialData))
 
-	testDecodeFloat(t, serialData, 0.1, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.3, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.5, MDK_RAW)
-	testDecodeFloat(t, serialData, 0.7, MDK_RAW)
-	requireFalse(t, canDecode(serialData), "Should not decode")
-}
-
-func TestSerialDataFull(t *testing.T) {
-	pl := []byte{0, 5, 92, 240, 74, 217, 134, 205, 44, 36, 83, 13, 63, 26,
-		62, 240, 68, 192, 41, 178, 7, 11, 166, 152, 233, 160, 61, 13,
-		225, 17, 145, 35, 33, 50, 159, 69, 190, 44}
-	serialData := NewSerialDataWithBuffer(pl)
-
-	Require(t, canDecode(serialData))
-
-	testDecodeFloat(t, serialData, 66.66, MDK_TEMP)
-	testDecodeFloat(t, serialData, 55.55, MDK_RH)
-	testDecodeFloat(t, serialData, 26262, MDK_LUX)
-	testDecodeFloat(t, serialData, 1, MDK_MOVEMENT)
-	testDecodeFloat(t, serialData, 199999, MDK_COUNTER)
-	testDecodeFloat(t, serialData, 1, MDK_DIGITAL)
-	testDecodeFloat(t, serialData, 8.88, MDK_VOLTAGE_0_10)
-	testDecodeFloat(t, serialData, 16.16, MDK_MILLIAMPS_4_20)
-	testDecodeFloat(t, serialData, 444444, MDK_OHM)
-	testDecodeFloat(t, serialData, 333, MDK_CO2)
-	testDecodeFloat(t, serialData, 2.9, MDK_BATTERY_VOLTAGE)
-	testDecodeFloat(t, serialData, 15, MDK_PUSH_FREQUENCY)
-	testDecodeFloat(t, serialData, 0.888, MDK_RAW)
-
-	testDecodeFloat(t, serialData, 0.1, MDK_UO)
-	testDecodeFloat(t, serialData, 0.2, MDK_UI)
-	testDecodeFloat(t, serialData, 0, MDK_DO)
-	testDecodeFloat(t, serialData, 1, MDK_DI)
-
-	testDecodeFloat(t, serialData, 22, MDK_FIRMWARE_VERSION)
-	testDecodeFloat(t, serialData, 44, MDK_HARDWARE_VERSION)
-
+	testDecodeFloat(t, serialData, 0.1, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.3, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.5, MDK_ANALOG_IN)
+	testDecodeFloat(t, serialData, 0.7, MDK_ANALOG_IN)
 	requireFalse(t, canDecode(serialData), "Should not decode")
 }